Remove secretProviderClass from leaflet-reader-deployment.yaml
The conditional block for secretProviderClass was removed from the leaflet-reader-deployment.yaml file. Now, the code only involves the "secret" block, making it more straightforward and focused mainly on the necessary secrets required for the deployment. The update also involves several changes like updating the chart's version and adjusting related documentation.