Update epi and app version, adjust config flags
This commit updates the epi chart to version 0.8.6, the app version to 3.0.0, and modifies several config variables. Most notably, development mode was switched off, the DSU fabric enclave type was modified to WalletDBEnclave, and the image repository was updated. The README instructions and code snippets were also updated to reflect these changes.